examples/python3/requirements.txt
paho-mqtt~=1.4
